+++ tank/minimap.h
@@ -16,6 +16,6 @@ extern bool badges;
 void setHeight(float h);
 float getHeight();
 void drawMiniMap(int ECRANW, int ECRANH, Tank& playerTank, EffectsContext& context, Camera& cam,
-    Type weather, bool badges, float dt);
+    Type weather, bool badges, double dt);
